Fixing bugs in frontend

+++ client/src/Components/Admin/ListProducts.js
@@ -11,7 +11,6 @@ export default function ListProducts(props) {
           AdminServices.products().then(data => {
                setProducts(data)
           })
-
      }, [])
 
      const delete_product = (id_product) => {
@@ -31,43 +30,53 @@ export default function ListProducts(props) {
      const edit_product = (id_product) => {
           history.push('/edit/'+id_product)
      }
-
-     return (
-          <div>
+     if(products.length === 0){
+          return(
+               <div className="container mx-auto text-center  mt-5">
+                    <h2>¡Aun no hay publicaciones!</h2>
+               </div>
+          )
+     }else{
+          return (
                <div>
-                    <h2>Productos publicados</h2>
-               </div>
-              <div className="container col-md-10 mx-auto mt-5">
-                    <table className="table">
-                         <thead>
-                              <tr className="text-center">
-                                   <th scope="col">Id</th>
-                                   <th scope="col">Titulo</th>
-                                   <th scope="col">Categoria</th>
-                                   <th scope="col">Precio</th>
-                                   <th scope="col" className="text-warning">Ver</th>
-                                   <th scope="col" className="text-primary">Editar</th>
-                                   <th scope="col" className="text-danger">Borrar</th>
-                              </tr>
-                         </thead>
-                         <tbody>
-                              {products.map(product => {
-                                   return(
-                                   <tr key={product.id_product} className="text-center">
-                                        <th scope="row">{product.id_product}</th>
-                                        <th scope="row">{product.title}</th>
-                                        <th scope="row">{product.categories}</th>
-                                        <th scope="row">${product.price}</th>
-                                        <th scope="row"><i style={{cursor:'pointer'}} onClick={() => view(product.id_product)} className="fa fa-plus"></i></th>
-                                        <th scope="row"><i style={{cursor:'pointer'}} onClick={() => edit_product(product.id_product)} className="fa fa-pencil"></i></th>
-                                        <th scope="row"><i style={{cursor:'pointer'}} onClick={() => delete_product(product.id_product)} className="fa fa-trash"></i></th>
+                    <div>
+                         <h2>Productos publicados</h2>
+                    </div>
+                   <div className="container col-md-10 mx-auto mt-5">
+                         <table className="table">
+                              <thead>
+                                   <tr className="text-center">
+                                        <th scope="col">Id</th>
+                                        <th scope="col">Titulo</th>
+                                        <th scope="col">Categoria</th>
+                                        <th scope="col">Precio</th>
+                                        <th scope="col" className="text-warning">Ver</th>
+                                        <th scope="col" className="text-primary">Editar</th>
+                                        <th scope="col" className="text-danger">Borrar</th>
                                    </tr>
-                                   ) 
-                              })}
-                              
-                         </tbody>
-                    </table>
+                              </thead>
+                               <tbody>
+                                   
+                                   {products.map(product => {
+                                        return(
+                                        <tr key={product.id_product} className="text-center">
+                                             <th scope="row">{product.id_product}</th>
+                                             <th scope="row">{product.title}</th>
+                                             <th scope="row">{product.categories}</th>
+                                             <th scope="row">${product.price}</th>
+                                             <th scope="row"><i style={{cursor:'pointer'}} onClick={() => view(product.id_product)} className="fa fa-plus"></i></th>
+                                             <th scope="row"><i style={{cursor:'pointer'}} onClick={() => edit_product(product.id_product)} className="fa fa-pencil"></i></th>
+                                             <th scope="row"><i style={{cursor:'pointer'}} onClick={() => delete_product(product.id_product)} className="fa fa-trash"></i></th>
+                                        </tr>
+                                        ) 
+                                   })}
+                                   
+                              </tbody> 
+                         </table>
+                    </div>
                </div>
-          </div>
-     )
+          )
+     }
+
+     
 }
